requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
numpy==1.20.3
pandas==1.2.4
gunicorn==20.0.4
pandas-profiling
scipy
sklearn==0.0
sklearn-pandas==1.8.0
SQLAlchemy==1.3.13
xgboost==0.90
zipp==2.0.1
catboost==1.0.3
boto3
statsmodels
imblearn
cassandra-driver